Refactoring neo4j.test.js for independent it()s (#143)
